% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manBetaCIWald-package.R
\docType{package}
\name{manBetaCIWald-package}
\alias{manBetaCIWald-package}
\alias{_PACKAGE}
\title{manBetaCIWald: betaDelta and betaSandwich: Confidence Intervals for Standardized Regression Coefficients in R}
\description{
Research compendium for the manuscript Pesigan, I. J. A., Sun, R. W., & Cheung, S. F. (2023). betaDelta and betaSandwich: Confidence intervals for standardized regression coefficients in R. Multivariate Behavioral Research. \doi{10.1080/00273171.2023.2201277}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/manBetaCIWald}
  \item \url{https://jeksterslab.github.io/manBetaCIWald/}
  \item \url{https://osf.io/8hy7w/}
  \item \doi{10.1080/00273171.2023.2201277}
  \item Report bugs at \url{https://github.com/jeksterslab/manBetaCIWald/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Rongwei Sun \email{irissun_s@hotmail.com} (\href{https://orcid.org/0000-0003-0034-1422}{ORCID}) [contributor]
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID}) [contributor]
}

}
\keyword{internal}
