% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manBetaCIWald-package.R
\docType{package}
\name{manBetaCIWald-package}
\alias{manBetaCIWald-package}
\alias{_PACKAGE}
\title{manBetaCIWald: betaDelta and betaSandwich: Confidence Intervals for Standardized Regression Coefficients in R}
\description{
Research compendium for the manuscript betaDelta and betaSandwich: Confidence Intervals for Standardized Regression Coefficients in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/manBetaCIWald}
  \item \url{https://jeksterslab.github.io/manBetaCIWald/}
  \item Report bugs at \url{https://github.com/jeksterslab/manBetaCIWald/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID}) [contributor]
  \item Rongwei Sun \email{irissun_s@hotmail.com} (\href{https://orcid.org/0000-0003-0034-1422}{ORCID}) [contributor]
}

}
\keyword{internal}
